Move identity registry functionality to identity package
