+++ identity/registry/endpoints.go
