+++ identity/registry/endpoints_test.go
